pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer - SEO analysis tool"
authors = [{ name = "Irina", email = "maxis2007@mail.ru" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.3",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
